chore: move `wasm_ctx`
