+++ src/pcd/mod.rs
@@ -13,8 +13,8 @@ use wasmi_wasi::WasiCtx;
 use crate::{
   circuits::{supernova::etable_rom::wasm_nivc, verify::verify_receipts},
   errors::ProvingError,
-  traits::args::ZKWASMContext,
   utils::nivc::build_rom,
+  wasm_ctx::ZKWASMContext,
 };
 
 use receipt::Receipt;
