+++ src/run/default/mod.rs
@@ -13,13 +13,13 @@ use crate::{
     supernova::etable_rom::EtableROM,
   },
   traits::{
-    args::ZKWASMContext,
     prover::Prover,
     public_values::{PublicValuesTrait, ZKVMPublicParams, ZKVMPublicValues},
     snark::RecursiveSNARKTrait,
     zkvm::{ZKVMBuilder, ZKVM},
   },
   utils::{nivc::build_rom, wasm::print_pretty_results},
+  wasm_ctx::ZKWASMContext,
 };
 use anyhow::anyhow;
 use ff::Field;
@@ -469,10 +469,8 @@ mod tests {
   };
 
   use crate::{
-    args::{WASMArgsBuilder, WASMCtx},
-    run::default::ZKEProof,
-    traits::zkvm::ZKVM,
-    utils::logging::init_logger,
+    args::WASMArgsBuilder, run::default::ZKEProof, traits::zkvm::ZKVM, utils::logging::init_logger,
+    wasm_ctx::WASMCtx,
   };
 
   type EE1<E> = ipa_pc::EvaluationEngine<E>;
