+++ src/run/tests.rs
@@ -1,9 +1,7 @@
 use std::path::PathBuf;
 
 use crate::{
-  args::{WASMArgsBuilder, WASMCtx},
-  traits::zkvm::ZKVM,
-  utils::logging::init_logger,
+  args::WASMArgsBuilder, traits::zkvm::ZKVM, utils::logging::init_logger, wasm_ctx::WASMCtx,
   BatchedZKEngine,
 };
